// src/lib/client/types/schemas.ts
export interface Community {
  id: CommunityId
  name: string
  title: string
  description?: string
  removed: boolean
  published: string
  updated?: string
  deleted: boolean
  nsfw: boolean
  actor_id: string
  local: boolean
  icon?: string
  banner?: string
  hidden: boolean
  posting_restricted_to_mods: boolean
  instance_id: InstanceId
  visibility: CommunityVisibility
}

export interface CommunityAggregates {
  community_id: CommunityId
  subscribers: number
  posts: number
  comments: number
  published: string
  users_active_day: number
  users_active_week: number
  users_active_month: number
  users_active_half_year: number
  subscribers_local: number
}

export interface CommunityBlockView {
  person: Person
  community: Community
}

export interface CommunityFollowerView {
  community: Community
  follower: Person
}
export type CommunityId = number

export interface CommunityModeratorView {
  community: Community
  moderator: Person
}

export interface CommunityView {
  community: Community
  subscribed: SubscribedType
  blocked: boolean
  counts: CommunityAggregates
  banned_from_community: boolean
}
export type CommunityVisibility = 'Public' | 'LocalOnly'
// ...
export interface Instance {
  id: InstanceId
  domain: string
  published: string
  updated?: string
  software?: string
  version?: string
}

export interface InstanceBlockView {
  person: Person
  instance: Instance
  site?: Site
}
export type InstanceId = number

export interface InstanceWithFederationState {
  id: InstanceId
  domain: string
  published: string
  updated?: string
  software?: string
  version?: string
  federation_state?: ReadableFederationState
}

export interface Language {
  id: LanguageId
  code: string
  name: string
}
export type LanguageId = number
// ...
export interface LocalUser {
  id: LocalUserId
  person_id: PersonId
  email?: string
  show_nsfw: boolean
  theme: string
  default_sort_type: SortType
  default_listing_type: ListingType
  interface_language: string
  show_avatars: boolean
  send_notifications_to_email: boolean
  show_scores: boolean
  show_bot_accounts: boolean
  show_read_posts: boolean
  email_verified: boolean
  accepted_application: boolean
  open_links_in_new_tab: boolean
  blur_nsfw: boolean
  auto_expand: boolean
  infinite_scroll_enabled: boolean
  admin: boolean
  post_listing_mode: PostListingMode
  totp_2fa_enabled: boolean
  enable_keyboard_navigation: boolean
  enable_animated_images: boolean
  collapse_bot_comments: boolean
  last_donation_notification: string
}
export type LocalUserId = number

export interface LocalUserView {
  local_user: LocalUser
  local_user_vote_display_mode: LocalUserVoteDisplayMode
  person: Person
  counts: PersonAggregates
}

export interface LocalUserVoteDisplayMode {
  local_user_id: LocalUserId
  score: boolean
  upvotes: boolean
  downvotes: boolean
  upvote_percentage: boolean
}
// ...
export interface MyUserInfo {
  local_user_view: LocalUserView
  follows: Array<CommunityFollowerView>
  moderates: Array<CommunityModeratorView>
  community_blocks: Array<CommunityBlockView>
  instance_blocks: Array<InstanceBlockView>
  person_blocks: Array<PersonBlockView>
  discussion_languages: Array<LanguageId>
}
export interface OpenGraphData {
  title?: string
  description?: string
  image?: string
  embed_video_url?: string
}
export type PaginationCursor = string
export interface PasswordChangeAfterReset {
  token: string
  password: string
  password_verify: string
}
export interface PasswordReset {
  email: string
}

export interface Person {
  id: PersonId
  name: string
  display_name?: string
  avatar?: string
  banned: boolean
  published: string
  updated?: string
  actor_id: string
  bio?: string
  local: boolean
  banner?: string
  deleted: boolean
  matrix_user_id?: string
  bot_account: boolean
  ban_expires?: string
  instance_id: InstanceId
}

export interface PersonAggregates {
  person_id: PersonId
  post_count: number
  comment_count: number
}

export interface PersonBlockView {
  person: Person
  target: Person
}
// ...
export interface Site {
  id: SiteId
  name: string
  sidebar?: string
  published: string
  updated?: string
  icon?: string
  banner?: string
  description?: string
  actor_id: string
  last_refreshed_at: string
  inbox_url: string
  public_key: string
  instance_id: InstanceId
  content_warning?: string
}